requirements.txt
requests
pyyaml
logging
influx_line_protocol
influxdb_client